requirements.txt
fastapi>=0.100.0
uvicorn[standard]==0.29.0
google-cloud-vision==3.7.4
google-generativeai==0.5.4
supabase==1.0.3
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic>=2.0.0
httpx
rsa==4.9
idna==3.9
openai
